Split diagnostic tests
